Refactor evaluation traits (#12)
Previously, evaluation was tightly coupled to the `Tape` type, and individual
evaluators were tightly bound to a specific tape.  Upon reflection, both of
these decisions were reconsidered:

- We may want to use Fidget's algorithms on implicit models that aren't defined
  as tapes, so the evaluation traits should be more generic
- The previous code put a great deal of effort into reusing evaluator
  workspaces, but it was hard; it turns out that it's much cleaner to have
  evaluators own their workspaces but **not** own their tapes

These changes can be seen in the new `fidget::eval::Shape` trait, which replaces
(but is not equivalent to) the previous `fidget::eval::Family` trait.

+++ fidget/src/mesh/mt/dc.rs
@@ -250,7 +250,7 @@ impl DcBuilder for DcWorker<'_> {
         let mut next = self.verts.len();
         assert!(next < (1 << 55));
         next |= 1 << 63;
-        next |= (self.thread_index as usize) << 55;
+        next |= self.thread_index << 55;
 
         match self.map[i].compare_exchange(
             0,
